+++ cumulusci/tasks/salesforce.py
@@ -11,6 +11,9 @@ from cumulusci.salesforce_api.metadata import ApiDeploy
 from cumulusci.salesforce_api.metadata import ApiRetrieveInstalledPackages
 from cumulusci.salesforce_api.metadata import ApiRetrievePackaged
 from cumulusci.salesforce_api.metadata import ApiRetrieveUnpackaged
+from cumulusci.salesforce_api.package_zip import CreatePackageZipBuilder
+from cumulusci.salesforce_api.package_zip import InstallPackageZipBuilder
+from cumulusci.salesforce_api.package_zip import UninstallPackageZipBuilder
 
 class BaseSalesforceTask(BaseTask):
     name = 'BaseSalesforceTask'
@@ -35,10 +38,7 @@ class BaseSalesforceMetadataApiTask(BaseSalesforceTask):
 
     def _run_task(self):
         api = self._get_api()
-        if self.options:
-            return api(**options)
-        else:
-            return api()
+        return api()
 
 class BaseSalesforceApiTask(BaseSalesforceTask):
     name = 'BaseSalesforceApiTask'
@@ -107,6 +107,67 @@ class Deploy(BaseSalesforceMetadataApiTask):
 
         return self.api_class(self, package_zip)
 
+class CreatePackage(Deploy):
+    task_options = {
+        'package': {
+            'description': 'The name of the package to create.  Defaults to project__package__name',
+            'required': True,
+        },
+        'api_version': {
+            'description': 'The api version to use when creating the package.  Defaults to project__package__api_version',
+            'required': True,
+        },
+    }
+
+    def _init_options(self, kwargs):
+        super(CreatePackage, self)._init_options(kwargs)
+        if 'package' not in self.options:
+            self.options['package'] = self.project_config.project__package__name
+        if 'api_version' not in self.options:
+            self.options['api_version'] = self.project_config.project__package__api_version
+
+    def _get_api(self):
+        package_zip = CreatePackageZipBuilder(self.options['package'], self.options['api_version'])
+        return self.api_class(self, package_zip())
+
+class InstallPackageVersion(Deploy):
+    task_options = {
+        'namespace': {
+            'description': 'The namespace of the package to install.  Defaults to project__package__namespace',
+            'required': True,
+        },
+        'version': {
+            'description': 'The version of the package to install',
+            'required': True,
+        },
+    }
+
+    def _init_options(self, kwargs):
+        super(InstallPackageVersion, self)._init_options(kwargs)
+        if 'namespace' not in self.options:
+            self.options['namespace'] = self.project_config.project__package__namespace
+
+    def _get_api(self):
+        package_zip = InstallPackageZipBuilder(self.options['namespace'], self.options['version'])
+        return self.api_class(self, package_zip())
+
+class UninstallPackage(Deploy):
+    task_options = {
+        'namespace': {
+            'description': 'The namespace of the package to uninstall.  Defaults to project__package__namespace',
+            'required': True,
+        },
+    }
+
+    def _init_options(self, kwargs):
+        super(UninstallPackage, self)._init_options(kwargs)
+        if 'namespace' not in self.options:
+            self.options['namespace'] = self.project_config.project__package__namespace
+
+    def _get_api(self):
+        package_zip = UninstallPackageZipBuilder(self.options['namespace'])
+        return self.api_class(self, package_zip())
+
 class DeployBundles(Deploy):
     task_options = {
         'path': {
@@ -138,3 +199,64 @@ class DeployBundles(Deploy):
             return api(**options)
         else:
             return api()
+
+class PackageUpload(BaseSalesforceToolingApiTask):
+    name = 'PackageUpload'
+    task_options = {
+        'name': {
+            'description': 'The name of the package version.',
+            'required': True,
+        },
+        'production': {
+            'description': 'If True, uploads a production release.  Defaults to uploading a beta',
+        },
+        'description': {
+            'description': 'A description of the package and what this version contains.',
+        },
+        'password': {
+            'description': "An optional password for sharing the package privately with anyone who has the password. Don't enter a password if you want to make the package available to anyone on AppExchange and share your package publicly.",
+        },
+        'post_install_url': {
+            'description': 'The fully-qualified URL of the post-installation instructions. Instructions are shown as a link after installation and are available from the package detail view.',
+        },
+        'release_notes_url': {
+            'description': 'The fully-qualified URL of the package release notes. Release notes are shown as a link during the installation process and are available from the package detail view after installation.',
+        },
+        'namespace': {
+            'description': 'The namespace of the package.  Defaults to project__package__namespace',
+        },
+    }
+
+    def _init_options(self, kwargs):
+        super(PackageUpload, self)._init_options(kwargs)
+
+        # Set the namespace option to the value from cumulusci.yml if not already set
+        if not 'namespace' in self.options:
+            self.options['namespace'] = self.project_config.project__package__namespace
+
+    def _run_task(self):
+        sf = self._init_api()
+        package_res = sf.query("select Id from MetadataPackage where NamespacePrefix=''".format(self.options['namespace']))
+
+        production = self.options.get('production', False) in [True, 'True', 'true']
+        package_info = {
+            'VersionName': self.options['name'],
+            'IsReleaseVersion': production,
+        }
+        
+        if 'description' in self.options:
+            package_info['Description'] = self.options['description']
+        if 'password' in self.options:
+            package_info['Password'] = self.options['password']
+        if 'post_install_url' in self.options:
+            package_info['PostInstallUrl'] = self.options['post_install_url']
+        if 'release_notes_url' in self.options:
+            package_info['ReleaseNotesUrl'] = self.options['release_notes_url']
+
+        upload = self.tooling.PackageUploadRequest.create(package_info)[0]
+
+        while upload.status == 'InProgress':
+            time.sleep(3)
+            upload = self.tooling.query("select Status, Errors from PackageUploadRequest where Id = '{}'".format(upload['Id']))[0]
+
+        self.logger.info('Uploaded package version: {}'.format(str(upload)))